src/main/java/com/bakhtiyor/spittr/controllers/SpittleController.java
package com.bakhtiyor.spittr.controllers;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import com.bakhtiyor.spittr.models.Spittle;
import com.bakhtiyor.spittr.repositories.SpittleRepository;

@Controller
@RequestMapping("/spittles")
public class SpittleController {

	private static final String LONG_MAX_AS_STRING = "9223372036854775807";

	private SpittleRepository spittleRepository;

	@Autowired
	public SpittleController(SpittleRepository spittleRepository) {
		this.spittleRepository = spittleRepository;
	}

	@RequestMapping(method = RequestMethod.GET)
	public List<Spittle> spittles(
		@RequestParam(value = "max", defaultValue = LONG_MAX_AS_STRING) Long max,
		@RequestParam(value = "count", defaultValue = "20") Integer count
	) {
		return spittleRepository.findSpittles(max, count);
	}

	@RequestMapping(value = "/{spittleId}", method = RequestMethod.GET)
	public String spittle(@PathVariable("spittleId") Long spittleId, Model model) {
		model.addAttribute(spittleRepository.findOne(spittleId));
		return "spittle";
	}
}